backend/requirements.txt
# Django core
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# CORS
django-cors-headers>=4.3,<5.0

# Filtering
django-filter>=24.0,<25.0

# Async tasks
celery>=5.4,<6.0
redis>=5.0,<6.0

# Image processing (avatars, diploma scans)
Pillow>=10.0,<11.0

# Environment variables
python-dotenv>=1.0,<2.0

# Production server
gunicorn>=22.0,<23.0
